refactor wasmple_bridge crates structure.
