+++ wasmple/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "wasmple"
 description = "a sample project for WebAssembly with Rust & SolidJS."
-version = "0.4.2"
+version = "0.4.3"
 edition = "2021"
 
 [lib]
@@ -11,7 +11,6 @@ crate-type = ["cdylib", "rlib"]
 itertools = "0.10.5"
 serde = { version = "1.0.147", features = ["derive"] }
 wasmple_bridge = { path = "../wasmple_bridge" }
-wasmple_bridge_impl = { path = "../wasmple_bridge_impl" }
 wasmple_buffer = { path = "../wasmple_buffer" }
 wasmple_console = { path = "../wasmple_console" }
